app/requirements/base.txt
gunicorn==20.1.0
requests==2.27.1
psycopg2-binary==2.9.2
mysqlclient==2.1.0
drf-yasg==1.20.0
djangorestframework==3.13.1
django-cors-headers==3.11.0
celery==5.2.3
redis==4.1.4
pandas==1.5.1
numpy==1.26.4
django-cookie-law==2.1.1
Django==4.0.2
python-dotenv
django-filter==21.1
django-celery-results==2.3.0
mssql-django==1.1.3
PyJWT
mysql-connector-python
oso
django-oso
